updated README
The basic login authentication is done and works as intended. Will be taking a look at other Authentication like Google or Facebook. Right now its setup with email and pass. Next step would be to get our location and list the nearby Restaurants (this can be changed later to be dynamic allocated?).